fix: file for PropertySelector
